misinformation_simulation/requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0

# Graph and Visualization
networkx>=3.0
matplotlib>=3.8.0

# UI
gradio>=4.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
